# openapi.yml
openapi: 3.0.3
info:
  title: HMS Backend API
  version: 1.0.0
  description: |
    OpenAPI specification based solely on the application's actual Express routes and validation rules.
  contact:
    name: HMS Admin
    email: mrfortrancer@gmail.com
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://hms.fortrancer.in
    description: Production server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request payload or missing required information.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication is required or token is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Caller does not have the required permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: Something went wrong.
        statusCode:
          type: integer
          example: 500
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
        details:
          type: object
          additionalProperties: true
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully.
...
    AppointmentCreateRequest:
      type: object
      required:
        - patientId
        - doctorEmployeeId
        - date
        - timeSlot
        - createdByEmployeeId
      properties:
        patientId:
          type: string
          example: PAT-1001
        doctorEmployeeId:
          type: string
          example: EMP-2001
        date:
          type: string
          example: 2026-07-15
        timeSlot:
          type: string
          example: 10:30
        status:
          type: string
          example: Booked
        createdByEmployeeId:
          type: string
          example: EMP-1001
    AppointmentEditRequest:
      type: object
      required:
        - appointmentId
        - patientId
        - doctorEmployeeId
        - date
        - timeSlot
      properties:
        appointmentId:
          type: string
          example: APT-9001
        patientId:
          type: string
          example: PAT-1001
        doctorEmployeeId:
          type: string
          example: EMP-2001
        date:
          type: string
          example: 2026-07-16
        timeSlot:
          type: string
          example: 11:00
        status:
          type: string
          example: Rescheduled
    AppointmentStatusRequest:
      type: object
      required:
        - appointmentId
        - status
      properties:
        appointmentId:
          type: string
          example: APT-9001
        status:
          type: string
          example: Completed
    AppointmentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
          example: 23
        page:
          type: integer
          example: 1
        totalPages:
          type: integer
          example: 5
...
paths:
...
  /appointment/createAppointment:
    post:
      tags: [Appointment]
      summary: Create appointment
      description: Schedules a new appointment.
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreateRequest'
      responses:
        '201':
          description: Appointment created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Appointment Created Successfully
                  date:
                    type: string
                  status:
                    type: string
                    example: Booked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointment/getAllAppointments:
    get:
      tags: [Appointment]
      summary: Get all appointments
      description: Returns paginated appointments.
      operationId: getAllAppointments
      parameters:
        - in: query
          name: page
          required: true
          schema:
            type: integer
            example: 1
        - in: query
          name: limit
          required: true
          schema:
            type: integer
            example: 10
        - in: query
          name: selectedText
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Appointments retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointment/getDoctors:
    get:
      tags: [Appointment]
      summary: Get doctors list
      description: Returns active doctors for appointment booking.
      operationId: getDoctors
      responses:
        '200':
          description: Doctors returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointment/getAppointmentUiData:
    get:
      tags: [Appointment]
      summary: Get appointment UI data
      description: Returns appointment dashboard summary counts.
      operationId: getAppointmentUiData
      responses:
        '200':
          description: Appointment UI data returned successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  appointmentCount:
                    type: integer
                  bookedCount:
                    type: integer
                  cancelledCount:
                    type: integer
                  completedCount:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointment/deleteAppointment:
    get:
      tags: [Appointment]
      summary: Delete appointment
      description: Soft-deletes an appointment by ID.
      operationId: deleteAppointment
      parameters:
        - in: query
          name: appointmentId
          required: true
          schema:
            type: string
            example: APT-9001
        - in: query
          name: deletedBy
          required: false
          schema:
            type: string
            example: EMP-1001
      responses:
        '200':
          description: Appointment deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointment/getAppointmentsByPatientId:
    get:
      tags: [Appointment]
      summary: Get appointments by patient
      description: Returns appointments filtered by patientId.
      operationId: getAppointmentsByPatientId
      parameters:
        - in: query
          name: patientId
          required: true
          schema:
            type: string
            example: PAT-1001
        - in: query
          name: selectedText
          required: false
          schema:
            type: string
        - in: query
          name: page
          required: true
          schema:
            type: integer
            example: 1
        - in: query
          name: limit
          required: true
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: Appointments returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointment/getDoctorByEmployeeId:
    get:
      tags: [Appointment]
      summary: Get doctor by employee ID
      description: Returns doctor data for the supplied employeeId.
      operationId: getDoctorByEmployeeId
      parameters:
        - in: query
          name: employeeId
          required: true
          schema:
            type: string
            example: EMP-2001
      responses:
        '200':
          description: Doctor returned successfully.
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointment/editAppointment:
    post:
      tags: [Appointment]
      summary: Edit appointment
      description: Updates appointment details by appointmentId.
      operationId: editAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentEditRequest'
      responses:
        '200':
          description: Appointment updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointment/editAppointmentStatus:
    post:
      tags: [Appointment]
      summary: Edit appointment status
      description: Updates only an appointment status.
      operationId: editAppointmentStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentStatusRequest'
      responses:
        '200':
          description: Appointment status updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /appointment/getAppointmentByDoctorIdOrPatientId:
    get:
      tags: [Appointment]
      summary: Get appointment by doctor, patient, or appointment ID
      description: Returns completed appointments matching supplied identifiers.
      operationId: getAppointmentByDoctorIdOrPatientId
      parameters:
        - in: query
          name: doctorId
          required: false
          schema:
            type: string
            example: EMP-2001
        - in: query
          name: patientId
          required: false
          schema:
            type: string
            example: PAT-1001
        - in: query
          name: appointmentId
          required: false
          schema:
            type: string
            example: APT-9001
      responses:
        '200':
          description: Appointments returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...